fix code educator, simulation, task, subtask, task question, student subtask progress, student task question progress

+++ source/user-base-auth/src/main/java/com/base/auth/controller/TaskController.java
@@ -16,6 +16,7 @@ import com.base.auth.model.SubTask;
 import com.base.auth.model.Task;
 import com.base.auth.model.criteria.TaskCriteria;
 import com.base.auth.repository.SimulationRepository;
+import com.base.auth.repository.StudentSubTaskProgressRepository;
 import com.base.auth.repository.SubTaskRepository;
 import com.base.auth.repository.TaskQuestionRepository;
 import com.base.auth.repository.TaskRepository;
@@ -59,6 +60,9 @@ public class TaskController extends ABasicController{
 
   @Autowired
   TaskQuestionRepository taskQuestionRepository;
+
+  @Autowired
+  StudentSubTaskProgressRepository studentSubTaskProgressRepository;
 
   @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
   @PreAuthorize("hasRole('TA_C')")
@@ -184,6 +188,9 @@ public class TaskController extends ABasicController{
     -> new NotFoundException("Simulation not found", ErrorCode.SIMULATION_ERROR_NOT_FOUND));
     if (!Objects.equals(simulation.getEducator().getId(), getCurrentUser())){
       throw new BadRequestException("Simulation cannot be deleted", ErrorCode.SIMULATION_ERROR_NOT_AUTHORIZED);
+    }
+    if (studentSubTaskProgressRepository.existsStudentSubTaskProgressByTaskId(id)){
+      throw new BadRequestException("Task cannot be deleted", ErrorCode.TASK_ERROR_NOT_DELETE);
     }
     taskQuestionRepository.deleteAllTaskQuestionByTaskId(id);
     subTaskRepository.deleteByTaskId(id);
